requirements.txt
ebaysdk
pyyaml
prettytable